Curriforge - Curriculum Planner/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
google-generativeai==0.8.0
pydantic[email]==2.9.0
python-dotenv==1.0.1
httpx==0.27.0
bcrypt==4.2.0
